Backend/requirements.txt
fastapi
uvicorn[standard]
pandas
sqlalchemy
langchain-groq
langchain
python-dotenv
psycopg2